+++ ee/tabby-ui/lib/tabby/fetcher.ts
@@ -16,6 +16,7 @@ interface FetcherOptions extends RequestInit {
     input: RequestInfo | URL,
     init?: RequestInit | undefined
   ) => Promise<Response>
+  errorHandler?: (response: Response) => any
 }
 
 export default async function authEnhancedFetch(
@@ -24,7 +25,7 @@ export default async function authEnhancedFetch(
 ): Promise<any> {
   const currentFetcher = options?.customFetch ?? window.fetch
 
-  if (willAuthError()) {
+  if (willAuthError(url)) {
     return tokenManagerInstance.refreshToken(doRefreshToken).then(res => {
       return requestWithAuth(url, options)
     })
@@ -44,7 +45,10 @@ export default async function authEnhancedFetch(
   }
 }
 
-function willAuthError() {
+function willAuthError(url: string) {
+  if (url.startsWith('/oauth/providers')) {
+    return false
+  }
   const accessToken = getAuthToken()?.accessToken
   if (accessToken) {
     // Check whether `token` JWT is expired
@@ -97,9 +101,15 @@ function requestWithAuth(url: string, options?: FetcherOptions) {
 
 function formatResponse(
   response: Response,
-  options?: Pick<FetcherOptions, 'responseFormat' | 'responseFormatter'>
+  options?: Pick<
+    FetcherOptions,
+    'responseFormat' | 'responseFormatter' | 'errorHandler'
+  >
 ) {
-  if (!response?.ok) return undefined
+  if (!response?.ok) {
+    if (options?.errorHandler) return options.errorHandler(response)
+    return undefined
+  }
   if (options?.responseFormatter) {
     return options.responseFormatter(response)
   }