+++ ee/tabby-webserver/src/oauth/google.rs
@@ -1,6 +1,6 @@
 use std::sync::Arc;
 
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use async_trait::async_trait;
 use serde::Deserialize;
 
@@ -65,9 +65,7 @@ impl GoogleClient {
         credential: OAuthCredential,
         redirect_uri: String,
     ) -> Result<GoogleOAuthResponse> {
-        let Some(client_secret) = credential.client_secret else {
-            return Err(anyhow!("No client_secret present"));
-        };
+        let client_secret = credential.client_secret;
 
         let params = [
             ("client_id", credential.client_id.as_str()),