+++ ee/tabby-webserver/src/schema/repository.rs
@@ -1,15 +1,21 @@
 use async_trait::async_trait;
 use juniper::{GraphQLObject, ID};
+use lazy_static::lazy_static;
+use regex::Regex;
 use validator::Validate;
 
 use super::{Context, Result};
 use crate::juniper::relay::NodeType;
 
+lazy_static! {
+    static ref REPOSITORY_NAME_REGEX: Regex = Regex::new("").unwrap();
+}
+
 #[derive(Validate)]
 pub struct CreateRepositoryInput {
     #[validate(regex(
         code = "name",
-        path = "tabby_common::config::REPOSITORY_NAME_REGEX",
+        path = "self::REPOSITORY_NAME_REGEX",
         message = "Invalid repository name"
     ))]
     pub name: String,