+++ ee/tabby-webserver/src/service/job.rs
@@ -10,31 +10,31 @@ use crate::schema::{
 
 #[async_trait]
 impl JobService for DbConn {
-    async fn create_job_run(&self, name: String) -> Result<ID> {
+    async fn start(&self, name: String) -> Result<ID> {
         Ok(self.create_job_run(name).await.map(|x| x.as_id())?)
     }
 
-    async fn update_job_stdout(&self, id: &ID, stdout: String) -> Result<()> {
+    async fn update_stdout(&self, id: &ID, stdout: String) -> Result<()> {
         self.update_job_stdout(id.as_rowid()?, stdout).await?;
         Ok(())
     }
 
-    async fn update_job_stderr(&self, id: &ID, stderr: String) -> Result<()> {
+    async fn update_stderr(&self, id: &ID, stderr: String) -> Result<()> {
         self.update_job_stderr(id.as_rowid()?, stderr).await?;
         Ok(())
     }
 
-    async fn complete_job_run(&self, id: &ID, exit_code: i32) -> Result<()> {
+    async fn complete(&self, id: &ID, exit_code: i32) -> Result<()> {
         self.update_job_status(id.as_rowid()?, exit_code).await?;
         Ok(())
     }
 
-    async fn finalize_stale_job_runs(&self) -> Result<()> {
+    async fn cleanup(&self) -> Result<()> {
         (self as &DbConn).finalize_stale_job_runs().await?;
         Ok(())
     }
 
-    async fn list_job_runs(
+    async fn list(
         &self,
         ids: Option<Vec<ID>>,
         jobs: Option<Vec<String>>,
@@ -57,7 +57,7 @@ impl JobService for DbConn {
             .collect())
     }
 
-    async fn compute_job_run_stats(&self, jobs: Option<Vec<String>>) -> Result<JobStats> {
+    async fn compute_stats(&self, jobs: Option<Vec<String>>) -> Result<JobStats> {
         let stats = (self as &DbConn).compute_job_stats(jobs).await?;
         Ok(JobStats {
             success: stats.success,
@@ -75,19 +75,12 @@ mod tests {
     async fn test_job_service() {
         let svc: Box<dyn JobService> = Box::new(DbConn::new_in_memory().await.unwrap());
 
-        let id = svc.create_job_run("test-job".to_owned()).await.unwrap();
-        svc.update_job_stdout(&id, "stdout".to_owned())
-            .await
-            .unwrap();
-        svc.update_job_stderr(&id, "stderr".to_owned())
-            .await
-            .unwrap();
-        svc.complete_job_run(&id, 0).await.unwrap();
+        let id = svc.start("test-job".to_owned()).await.unwrap();
+        svc.update_stdout(&id, "stdout".to_owned()).await.unwrap();
+        svc.update_stderr(&id, "stderr".to_owned()).await.unwrap();
+        svc.complete(&id, 0).await.unwrap();
 
-        let job = svc
-            .list_job_runs(None, None, None, None, None, None)
-            .await
-            .unwrap();
+        let job = svc.list(None, None, None, None, None, None).await.unwrap();
         let job = job.first().unwrap();
         assert_eq!(job.job, "test-job");
         assert_eq!(job.stdout, "stdout");
@@ -95,14 +88,14 @@ mod tests {
         assert_eq!(job.exit_code, Some(0));
 
         let jobs = svc
-            .list_job_runs(Some(vec![id]), None, None, None, None, None)
+            .list(Some(vec![id]), None, None, None, None, None)
             .await
             .unwrap();
         assert_eq!(jobs.len(), 1);
 
-        svc.create_job_run("another-job".into()).await.unwrap();
+        svc.start("another-job".into()).await.unwrap();
         let jobs = svc
-            .list_job_runs(
+            .list(
                 None,
                 Some(vec!["another-job".into()]),
                 None,
@@ -120,22 +113,22 @@ mod tests {
         let db = DbConn::new_in_memory().await.unwrap();
         let jobs: Box<dyn JobService> = Box::new(db);
 
-        let id = jobs.create_job_run("test-job".into()).await.unwrap();
-        jobs.complete_job_run(&id, 0).await.unwrap();
+        let id = jobs.start("test-job".into()).await.unwrap();
+        jobs.complete(&id, 0).await.unwrap();
 
-        let id2 = jobs.create_job_run("test-job".into()).await.unwrap();
-        jobs.complete_job_run(&id2, 1).await.unwrap();
+        let id2 = jobs.start("test-job".into()).await.unwrap();
+        jobs.complete(&id2, 1).await.unwrap();
 
-        jobs.create_job_run("pending-job".into()).await.unwrap();
+        jobs.start("pending-job".into()).await.unwrap();
 
-        let stats = jobs.compute_job_run_stats(None).await.unwrap();
+        let stats = jobs.compute_stats(None).await.unwrap();
 
         assert_eq!(stats.success, 1);
         assert_eq!(stats.failed, 1);
         assert_eq!(stats.pending, 1);
 
         let stats = jobs
-            .compute_job_run_stats(Some(vec!["test-job".into()]))
+            .compute_stats(Some(vec!["test-job".into()]))
             .await
             .unwrap();