+++ ee/tabby-webserver/src/service/license.rs
@@ -6,9 +6,12 @@ use lazy_static::lazy_static;
 use serde::Deserialize;
 use tabby_db::DbConn;
 
-use crate::schema::{
-    license::{LicenseInfo, LicenseService, LicenseStatus, LicenseType},
-    Result,
+use crate::{
+    env::demo_mode,
+    schema::{
+        license::{LicenseInfo, LicenseService, LicenseStatus, LicenseType},
+        Result,
+    },
 };
 
 lazy_static! {
@@ -82,6 +85,18 @@ impl LicenseServiceImpl {
         }
         .guard_seat_limit())
     }
+
+    async fn make_demo_license(&self) -> Result<LicenseInfo> {
+        let seats_used = self.db.count_active_users().await? as i32;
+        Ok(LicenseInfo {
+            r#type: LicenseType::Enterprise,
+            status: LicenseStatus::Ok,
+            seats: 100,
+            seats_used,
+            issued_at: None,
+            expires_at: None,
+        })
+    }
 }
 
 pub async fn new_license_service(db: DbConn) -> Result<impl LicenseService> {
@@ -114,7 +129,11 @@ fn license_info_from_raw(raw: LicenseJWTPayload, seats_used: usize) -> Result<Li
 
 #[async_trait]
 impl LicenseService for LicenseServiceImpl {
-    async fn read_license(&self) -> Result<LicenseInfo> {
+    async fn read(&self) -> Result<LicenseInfo> {
+        if demo_mode() {
+            return self.make_demo_license().await;
+        }
+
         let Some(license) = self.db.read_enterprise_license().await? else {
             return self.make_community_license().await;
         };
@@ -126,7 +145,11 @@ impl LicenseService for LicenseServiceImpl {
         Ok(license)
     }
 
-    async fn update_license(&self, license: String) -> Result<()> {
+    async fn update(&self, license: String) -> Result<()> {
+        if demo_mode() {
+            return Err(anyhow!("Demo mode is enabled, cannot set license").into());
+        }
+
         let raw = validate_license(&license).map_err(|_e| anyhow!("License is not valid"))?;
         let seats = self.db.count_active_users().await?;
         match license_info_from_raw(raw, seats)?.status {
@@ -139,7 +162,7 @@ impl LicenseService for LicenseServiceImpl {
         Ok(())
     }
 
-    async fn reset_license(&self) -> Result<()> {
+    async fn reset(&self) -> Result<()> {
         self.db.update_enterprise_license(None).await?;
         Ok(())
     }
@@ -188,16 +211,16 @@ mod tests {
         let db = DbConn::new_in_memory().await.unwrap();
         let service = new_license_service(db).await.unwrap();
 
-        assert!(service.update_license("bad_token".into()).await.is_err());
+        assert!(service.update("bad_token".into()).await.is_err());
 
-        service.update_license(VALID_TOKEN.into()).await.unwrap();
-        assert!(service.read_license().await.is_ok());
+        service.update(VALID_TOKEN.into()).await.unwrap();
+        assert!(service.read().await.is_ok());
 
-        assert!(service.update_license(EXPIRED_TOKEN.into()).await.is_err());
+        assert!(service.update(EXPIRED_TOKEN.into()).await.is_err());
 
-        service.reset_license().await.unwrap();
+        service.reset().await.unwrap();
         assert_eq!(
-            service.read_license().await.unwrap().seats,
+            service.read().await.unwrap().seats,
             LicenseInfo::seat_limits_for_community_license() as i32
         );
     }